Refactor list page : 불필요한 코드 제거 및 클래스 네임 변경
