+++ src/pages/ProductList/Component/ListBottom/ListBottom.scss
@@ -1,10 +1,7 @@
 @import "../../../../styles/variables";
 
 div.listBottom {
-  .bottomTop {
-    // display: flex;
-    // align-items: end;
-    // justify-content: space-between;
+  .pagination {
     padding: 60px 0 15px 35px;
     border-bottom: 2px solid black;
 
@@ -23,18 +20,6 @@ div.listBottom {
       }
     }
 
-    .goodsCategory {
-      display: flex;
-      justify-content: space-between;
-      width: 250px;
-      padding-right: 20px;
-
-      button:hover {
-        opacity: 70%;
-        color: $main-grey;
-      }
-    }
-
     .fourButton {
       width: 25px;
       height: 25px;
@@ -42,10 +27,10 @@ div.listBottom {
   }
 }
 
-.bottomBottom {
+.itemList {
   display: flex;
 
-  .bottomLeft {
+  .listSorting {
     width: 20%;
     margin-top: 2%;
     margin-bottom: 20%;
@@ -53,13 +38,19 @@ div.listBottom {
     border-right: 1px solid black;
   }
 
-  .bottomRight {
+  .listMain {
     display: flex;
     flex-direction: column;
-    // gap: 50px 50px;
 
     width: 80%;
     padding: 5% 10%;
+
+    .artworkBox {
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: space-between;
+      width: 100%;
+    }
 
     .pageButton {
       display: flex;
@@ -71,12 +62,5 @@ div.listBottom {
         border: 1px solid black;
       }
     }
-  }
-
-  .artworkBox {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    width: 100%;
   }
 }
